+++ app/helpers/units_helper.rb
@@ -8,7 +8,7 @@ module UnitsHelper
   end
 
   def unit_actions(unit)
-    [
+    actions = [
       {
         label: I18n.t("ui.edit"),
         url: edit_unit_path(unit)
@@ -17,20 +17,26 @@ module UnitsHelper
         label: I18n.t("units.buttons.pdf_report"),
         url: report_unit_path(unit),
         target: "_blank"
-      },
-      {
+      }
+    ]
+
+    if unit.deletable?
+      actions << {
         label: I18n.t("units.buttons.delete"),
         url: unit,
         method: :delete,
         danger: true
-      },
-      {
-        label: I18n.t("units.buttons.add_inspection"),
-        url: inspections_path,
-        method: :post,
-        params: {unit_id: unit.id},
-        confirm: I18n.t("units.messages.add_inspection_confirm")
       }
-    ]
+    end
+
+    actions << {
+      label: I18n.t("units.buttons.add_inspection"),
+      url: inspections_path,
+      method: :post,
+      params: {unit_id: unit.id},
+      confirm: I18n.t("units.messages.add_inspection_confirm")
+    }
+
+    actions
   end
 end
